build row components
